fix reviewed issues 2
