+++ src/components/GenresFilter/GenresFilter.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { GENRE_FILTERS } from '@src/utils/constants';
-import { GenreQueries } from '@src/types/';
+import { GenreQueries, RootState } from '@src/types/';
 import { connect } from 'react-redux';
 import { fetchMovies } from '@src/store/actionCreators/movies';
-import { RootState } from '@src/store';
+import { selectGenre } from '@src/store/selectors/movies.selectors';
 import { GenreFilterProps } from './GenresFilter.types';
 import styles from './GenresFilter.module.scss';
 
@@ -31,11 +31,7 @@ const GenresFilter = ({ selected, filterMovies }: GenreFilterProps) => {
   );
 };
 
-const mapStateToProps = ({
-  movies: {
-    requestParams: { genre: selected },
-  },
-}: RootState) => ({ selected });
+const mapStateToProps = (state: RootState) => ({ selected: selectGenre(state) });
 
 const mapDispatchToProps = { filterMovies: fetchMovies };
