+++ src/context/app.context.ts
@@ -1,12 +1,13 @@
+import { noop } from '@src/utils/constants';
 import { createContext } from 'react';
 import { Context } from './context.types';
 
 const AppContext = createContext<Context>({
-  setActiveMovieId: () => {},
-  setEditingMovieId: () => {},
-  setShowAdd: () => {},
-  setShowEdit: () => {},
-  setShowDelete: () => {},
+  setActiveMovieId: noop,
+  setEditingMovieId: noop,
+  setShowAdd: noop,
+  setShowEdit: noop,
+  setShowDelete: noop,
 });
 
 export default AppContext;
