+++ src/store/actionCreators/movies.ts
@@ -1,15 +1,14 @@
 import API from '@src/api/api';
-import { Movie, RequestParams } from '@src/types';
+import { Movie, RequestParams, RootState, MoviesAction, MoviesActionType } from '@src/types';
 import { Dispatch } from 'redux';
-import { RootState } from '..';
-import { MoviesAction, MoviesActionType } from '../reducers/movies.reducer.types';
+import { selectRequestParams, selectMovies } from '../selectors/movies.selectors';
 
 export const fetchMovies =
   (params?: Partial<RequestParams>) =>
   (dispatch: Dispatch<MoviesAction>, getState: () => RootState) => {
     dispatch({ type: MoviesActionType.FETCH_MOVIES, payload: params });
 
-    API.getAll(getState().movies.requestParams)
+    API.getAll(selectRequestParams(getState()))
       .then((movies) => dispatch({ type: MoviesActionType.FETCH_MOVIES_SUCCESS, payload: movies }))
       .catch(() => dispatch({ type: MoviesActionType.FETCH_MOVIES_ERROR }));
   };
@@ -19,12 +18,23 @@ export const createMovie = (movie: Movie) => ({
   payload: movie,
 });
 
-export const updateMovie = (movie: Movie) => ({
-  type: MoviesActionType.UPDATE_MOVIE,
-  payload: movie,
-});
+export const updateMovie =
+  (movie: Movie) => (dispatch: Dispatch<MoviesAction>, getState: () => RootState) => {
+    const movies = selectMovies(getState());
+    const index = movies.findIndex(({ id }) => id === movie.id);
 
-export const deleteMovie = (id: number) => ({
-  type: MoviesActionType.DELETE_MOVIE,
-  payload: id,
-});
+    dispatch({
+      type: MoviesActionType.UPDATE_MOVIE,
+      payload: index < 0 ? movies : [...movies.slice(0, index), movie, ...movies.slice(index + 1)],
+    });
+  };
+
+export const deleteMovie =
+  (id: number) => (dispatch: Dispatch<MoviesAction>, getState: () => RootState) => {
+    const movies = selectMovies(getState());
+
+    dispatch({
+      type: MoviesActionType.DELETE_MOVIE,
+      payload: movies.filter((movie) => movie.id !== id),
+    });
+  };
