+++ src/store/index.ts
@@ -8,5 +8,4 @@ const rootReducer = combineReducers({
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
 
-export type RootState = ReturnType<typeof store.getState>;
 export default store;
